Java/src/main/java/arrays_and_hashing/valid_anagram/Solution.java
package arrays_and_hashing.valid_anagram;

import java.util.Arrays;

public class Solution {
    public boolean isAnagram(String s, String t) {
        if (s.length() != t.length()) {
            return false;
        }
        int[] count = new int[26];
        for (int i = 0; i < s.length(); i++) {
            int index = s.charAt(i) - 'a';
            count[index]++;
        }
        for (int i = 0; i < t.length(); i++) {
            int index = t.charAt(i) - 'a';
            count[index]--;
        }
        for (int value : count) {
            if (value != 0)
                return false;
        }
        return true;
    }
}
